[STYLE] Padding Label common/UI 이동
